Fix setting the grabbing mode
We need to queue this when setting it (as this comes from the main thread),
not when disabling it from within the render thread.

This also fixes cases where due to constant 60fps render pressure we would
actually not get around to process the queued disabling call, and thus pile
up captured frames in the event queue towards the main thread.