Thread safe TimerModel implementation
This implement thread safe handling for multi thread QTimer/QQmlTimer based timers.

This implement thread safe handling for multi thread free timers.

This implement automatic removal of invalid free timers.
As there is no way to know if a free timer is still running, free timers
those id is not reused by any other QTimer id will be considered valid
and not removed.

This implement automatic cleaning of obsolete history of free timers.
This happen when a free timer is stopped and a new one is started
reusing a previously used id.

Show free timers interval in the state column

Task-Id: KDEND-158