Replace hardcoded property return types with template type deduction
Easier to maintain, especially for properties changing return types over
time, where we can now avoid #ifdefs. This would be even nicer if MSVC2010
and 2012 would do the type deduction for overloaded member functions
correctly. So as long as we have to support those two, we need workarounds
for a few affected properties.