Update property overrides to handle multiple frontend instances
Multiple frontend instances will talk to the same backend plugin
instances. Users will expect the model to show the values once
rather have duplicates for all the instances.

This creates an overrider object per backend instance and manages
the list of front end instances. Changes to values from the plugin
UI are propagated to all the instances (which is needed if overriding
is one, in which case values don’t make it to the backend).

Also fix crash on shutdown with models not being cleaned up properly.

This depends on QtIVI change I2d756106c14536936451f61a502a13a1bb28e8dd