Add a way to query timers status update
We had to find a way to update the timer state when it changed
but we did not known it.
The chosen approach is to check the event dispatcher timers for a
changing object. This allow us check all the timers of a given dispatcher
at once still respecting the thread affinity.
QQmlTimer's use their runningChanged signal for that as they do not own
a timer id.

This also fix a OR check issue in canHandleCaller()

Task-Id: KDEND-224