Add support for down-casting to the meta object system
This is supposed to replace the various hard-coded down-casting cascades
we have in a number of places that deal with non-QObject instances. We
have the hierarchy information in the meta object data after all, no need
to maintain that twice. Additionally, it will just work in all places
encountering such types in the property model.

The test subject is QGradient, which turns out to be a pseudo-polymorphic
type. So, we need support for that too, which is why we have the type trait
to provide specialized dynamic_cast implementations in such a case.