Port server side decoration state to an auto-synced property
The manual handling was going out of sync if the client applied it's local
state to the server before that had created the overlay instance. This
doesn't happen with the property approach as the state can be stored now
even if the corresponding overlay is not yet existing or destroyed and
re-created.