Add property filter and adaptor for anchors.
Filter out all occurances of "anchors" from QMetaPropertyAdaptor and add
a custom property adaptor that uses `QQuickItemPrivate::_anchors`
instead of QMetaProperty::get. This is because the anchors-object is
created on-demand, but we don't want GammaRay to trigger the creation.
Instead we want to be able to show, whether the anchors-object exists
(in an unattached version of the program).

This adaptor is not yet able to notify about changes.