Adapt example and docs to the improved batcher
Since Qt 5.8 the batcher is able to batch rectangles that differ in
color, have rotations applied or have opacity <1 even if overlapping
(as long as all have the same opacity).
So with this patch the docs don't spread fake news about batching any
longer. Also the example was adapted to use different opacity values for
all the rectangles to trigger the unmerged batch performance issue.

Fixes #498