[WIP] Add ProblemReporter infrastructure
It allows to add or remove problems, is able to show them and offer
object and source navigation. It filters for duplicates and contains
some experimental problem providers, which are far from being
finished though.

The BindingInspector's problem provider doesn't work at all, even though
the code to look for and report the loops is in BindingModel::data and
thus definitely late enough. As the client shows it in the binding-tab,
it should definitevely be called...