Move binding inspector logic out of BindingModel
So far the logic to query BindingProviders and assemble the results,
building the binding tree structure, was part of the binding model.
Taken by itself, that's no good design. But it's indispensible to have
the logic available on its own as soon as we want to scan through all
objects to find binding loops.