Move binding aggregation logic even farther out
Recently, I moved the logic aggregating bindings from the various
binding providers out of the binding model into the binding extension.
Now this moves the stateless parts of that logic out of there into free
functions available from core/ directly.

This fixes the issue of the binding loop scan being triggered multiple times
from different tools, as it's no longer launched from the binding
extension but from the object inspector tool.

It has the additional advantage that now the static method to register a
binding provider has a proper place to stay, instead of requiring
plugins to include an implementation-header.