Allow filtering the problem list by source
This commit adds a list of available problem checkers that can be used
to filter the problem list by source and allows to preselect the scans
to be done before starting a scan.

To make this possible, the structure of the scans was refactored on the
server side: Instead of reacting to a signal we now register a callback,
which the ProblemCollector calls if the check is enabled. The custom
callback gives the opportunity to give the checkers a name and a
description and gives the ProblemCollector the power to choose which
ones to call.