Widget3DModel: replace broken pointer → string conversion with QString::asprintf()
The old code used memcpy to copy eight bytes from a uintptr_t to a QChar[8].
But the latter is 16 bytes, so we end up with all kinds of weird side-effects
in the string, like invalid characters or embedded NULs, or unmatched surrogate pairs.

Just use asprintf() to get a hex representation of the pointer.