Add probe null checks to signal spies
The destructor of the probe removes the signal spies.
But, QObject::doActivate() loads the signal spies, calls the thread and
then calls signal_end_callback. As the deletion of the probe itself is
triggered by a signal, signal_end_callback will be called right after
the deletion of the Probe, causing GammaRay to crash on exit.