Qt 5.14: Use pointer to signal spy set, not a copy
For Qt 5.14 qt_register_signal_spy_callbacks takes a pointer, not a
value anymore. Still, the probe copied the callback struct and on
destruction would call qt_register_signal_spy_callbacks with a pointer
to the function-local object, leading to a crash. Now, for Qt 5.14, we
don't copy the callback struct, but instead just store a copy of the
pointer and on Probe-destruction reset the qt_signal_spy_callback_set
with the old pointer value.