Allow capturing response for network operations
The Network Operations tab now contains a checkbox that allows the user
to enable collection of the response body for network replies.

To begin with, we only only collect up to 5MiB of data. In the future we
might make this configurable, and maybe even allow the user to whitelist
to specific Content-Type values.