Fix build when using Qt < 5.15.0
The support for capturing network response fiddles with the signal-slot
connections in QNetworkReply. The private API for this is different in
older Qt versions and hence the prioritizeLatestConnection function will
need to be adapted for those.
For now, disable the feature on older versions so that GammaRay can
still build. The function will then be a no-op which is fine because we
also hide the feature in such versions.