Guard current texture with smart QPointer
Hotfixes a scenario where a change of texture objects in a Texture view
could crash the GammaRay injector.

Smart QPointer to a QSGTexture is what a QSGTextureGrabber already uses,
which seems to work fine for it.

I noticed that it does not update textures in real-time as an inspected
button itself changes on a screen; but at least we don't segfault on it
anymore either.

Closes: #678