WidgetInspector: Auto expand the tree on filtering
When you are filtering for a particular widget, Gammaray doesn't expand
the tree for you. This can be really annoying when you have long object
hierarchies as you need to manually expand each node to get to the item
you want. To fix this, auto expand the tree after filtering.