Implement Object Favorite-ing
Gammaray shows huge lists of object usually but as a user, a person
is often only interest in a small set of objects. It becomes really
hard to switch between two objects.

Favorite-ing aims to make it easier. You can mark an object as favorite
and it will appear in a view above the current itemview. However, this
needs to be added to each itemview individually. This commit only
implements it for the Signals history view.