Try seeking the stream to the correct position on deserialization fail
If we get some data that can't be deserialized e.g., a User type that
somehow made it to the client, all subsequent data fails to get
deserialized and gammaray can crash.

To avoid this add a marker inside the data to restore reading from the
stream in case we get something bad.

This is especially important with Qt6.4 because in Qt6 if a type defines
QDatastream operator<< then our RemoteModel will happily send it across
to the client.