Fix invokeMethod failure with alias param
Starting from Qt 6.5, invokeMethod has changed a lot. For normal users
this doesn't matter too much, but for us, who still use the
QGenericArgument way it breaks things. InvokeMethod with QGenericArg no
longer checks aliases of a given argument and thus fails to call the
method.

To work around this, use the actual types in the signals instead.

The QGenericArgument overload of invokeMethod is marked for removal in
Qt7, so we might need to do it by hand in future but for now this will
work.