Fix QQuickWidget being shown incorrect in quick live view
With Qt 6.4, the content of widgets + quick seems to be on the same
surface. Thus if we start reading at 0,0 we end up showing widgets
and other unrelated stuff instead of the QQuickWidget that we are
supposed to show.

To fix this, calculate the offset of the QQuickWidget inside the
viewport.

#815