Don't mix up "custom command" and "system default" code navigation
When checking for the "CodeNavigation" key, GammaRay checks an int
variable in the QSettings. This is an index into an array which contains
editor for a platform, for example Linux has "Kate" as number 1 and so
on. The default value for this setting is "-1" which means "custom
command", but that's empty by default.

This results in nothing actually launching with the system default
editor on Linux, because there is no integer you can via in the UI
that's higher than 0, but not -1. Worse, this means that no text editor
is actually launched unless you set one yourself!

Now there's a better distinction between a "System Default" and the
pre-existing "Custom command" setting. Custom command is now -2, which
is still an invalid index into the ideSettings array so there's no
chance for collision still. However, the default -1 value is now "System
default" which improves the out-of-box experience dramatically.