Show classnames next to objectNames, when different
  Property     | Value                        | Type     | Class
Before:
  parent       | m_loadingWidget              | QObject* | QObject
After:
  parent       | m_loadingWidget (QSvgWidget) | QObject* | QObject

For consistency, change the case without object name from
"QFrame[this=0xdcd22e0]" to "0xdcd22e0 (QFrame)", as suggested
by ferdnyc.

The change propagated to other classes to fix const [in]correctness
(the const_cast is only because of QHash API, it shouldn't be necessary
for the caller of those methods).