Docs: Make QDOC_INDEX_DIR optional
If the Qt documentation index files are not available (which they
may not be, since some distro installs of Qt6 don't package the
documentation), the GammaRay manual can still be built. All of
its links to the system docs will be broken, but that doesn't
render it unusable.

Allow building the documentation even without access to the .index
files for the system docs, but print a message about broken links.