Unbreak recursive filtering in ObjectIdsFilterProxyModel and more
The code tried to be compatible with older Qt and newer but failed
to do so correctly. When a model only implemented `acceptRow` then
that was never called by newer Qt.

Instead, enforce all models to override `acceptRow` and prevent them
from overriding `filterAcceptsRow` - instead add the dance there to
make the code compatible to the old KRecursiveFilterProxyModel.

This unbreaks the "ctrl+shift+click" picker dialog to not show all
objects but really only those that are visible at the given position.