Add special check for libqtqmlcoreplugin when detecting Qt6 ABI
Qt6 has added a new QtCore import, which trips up our previous detection
of valid injectable processes. The QML plugin is called
"libqtqmlcoreplugin" and will take the place of a genuine QtCore library
we're looking for in a process.

This library isn't versioned though, so it's impossible to determine the
Qt version from the filename. This prevents it from being selected which
makes the process detection much more reliable for Qt6 QtQuick
applications.