First draft of RHI support
First steps to work on RHI. Tested on Linux/X11/Vulkan.

Drawing target decorations: unlike the OpenGL path, there isn't a
QRhiPaintDevice yet. So I paint decorations on a QImage, upload it as a
texture, and then use it to draw a fullscreen quad.

There's a flaw in the process, which is, the screen is captured *with*
the overlay drawn. Instead it should be captured without it.  It also
flickers a lot. I'm not sure what's the deal with the
grabbing/ungrabbing scheme that other captures have, it seems that we
get a frame without decorations and one with them. The relative code
looks also thread unsafe.

Missing:
* handling window resizes
* in general handling loss of RHI context, pipeline re-init, etc.
* handling QQuickWidget / QQuickRenderControl (completely untested. QQRC
  might just not work at all, as we don't have access to the swap chain
  in that case)
* reducing copies/allocations, a framegrab always allocates memory, plus
  the image for the painting the decorations on the target
  (assuming we don't find a better way)
* general threading fixup...