CPack: Fix icon paths
Some of those resources have never existed in the repository. But I
replaced them with TODO instead of removing entirely in a hope that
maybe someday someone would implement/draw them.

This patch fixes CPack enough to get it through the package and
installation stages. However, Qt dynamic libs are still missing,
so the installed exe can't launch as is. Needs windeployqt or something.