Fix infinite in loop QuickSceneGraphModel::itemForSgNode()
Window selection triggers lots of model and selection updates, which
are not atomic. It can happen that itemForSgNode() is called with
the old selected node from the old window, since it wasn't updated yet.

We could also attempt to prevent it from being called with old
sg node, but it's benign, since SG selection is processed right after.

For issue #1070