Fix GammaRay GUI not showing up on Windows
The target sends the TCP address via a QLocalSocket, but Windows
is very prone to partial writes.

Ensure all bytes are written before closing the socket.

Simplified the launcher code, no need to use a while, as the target
only sends 1 message. Ignore the readyRead that contain partial writes.

Related to some comments in issue #959